+++ test.js
@@ -1,33 +1,36 @@
-const fs = require('.');
+// tooling
+const fs   = require('.');
+const path = require('path');
 
+// tests
 Promise.all([
-	// test fs.readFile, fs.join
+	// test fs.readFile
 	fs.readFile(
-		fs.join(__dirname, 'package.json'),
+		path.join(__dirname, 'package.json'),
 		'utf-8'
 	).then(
 		(contents) => JSON.parse(contents)
 	).then(
 		(pkg) => pkg.name === 'fse'
 	),
-	// test fs.lstat, fs.mkdir, fs.join, fs.rmdir
+	// test fs.lstat, fs.mkdir, fs.rmdir
 	fs.lstat(
-		fs.join(__dirname, 'test')
+		path.join(__dirname, 'test')
 	).then(
 		() => Promise.reject(),
 		() => fs.mkdir(
-			fs.join(__dirname, 'test', 'quick', 'brown', 'fox')
+			path.join(__dirname, 'test', 'quick', 'brown', 'fox')
 		).then(
 			() => fs.lstat(
-				fs.join(__dirname, 'test')
+				path.join(__dirname, 'test')
 			)
 		).then(
 			() => fs.rmdir(
-				fs.join(__dirname, 'test')
+				path.join(__dirname, 'test')
 			)
 		).then(
 			() => fs.lstat(
-				fs.join(__dirname, 'test')
+				path.join(__dirname, 'test')
 			).catch(
 				(error) => error.code === 'ENOENT'
 			)