+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to fse
 
+### 2.0.0 (April 17, 2017)
+
+- Removed: watch task — sticking with plain fs promise-ified
+- Removed: path forwarding — sticking with plain fs promise-ified
+
 ### 1.0.1 (January 18, 2017)
 
 - Removed: Unnecessary catch in copyFile